Synchronise the compact view handling via config settings
