--- frontend/app/version.ts
@@ -1,11 +1,0 @@
-// VERSION_MAJOR is the semantic major version for the browser SPA runtime.
-export const VERSION_MAJOR = 1
-
-// VERSION_MINOR is the semantic minor version for the browser SPA runtime.
-export const VERSION_MINOR = 1
-
-// VERSION_PATCH is the semantic patch version for the browser SPA runtime.
-export const VERSION_PATCH = 0
-
-// APP_VERSION is the semantic version rendered in browser metadata and the page footer.
-export const APP_VERSION = `${VERSION_MAJOR}.${VERSION_MINOR}.${VERSION_PATCH}`
