+++ public/agents.html
@@ -194,8 +194,7 @@
       </footer>
     </main>
 
-    <script src="./js/page-meta.min.js"></script>
-    <script src="./js/top-menu.min.js"></script>
+    <script src="./js/page-chrome.min.js"></script>
     <script src="./js/tapoo.min.js"></script>
   </body>
 </html>
